requirements.txt
Flask
Pillow
torch
diffusers
accelerate
transformers